+++ lib/src/geojson.dart
@@ -3,11 +3,12 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:geodesy/geodesy.dart';
-import 'package:iso/iso.dart';
 import 'package:pedantic/pedantic.dart';
 
 import 'deserializers.dart';
 import 'exceptions.dart';
+import 'iso/iso.dart';
+import 'iso/runner.dart';
 import 'models.dart';
 
 /// The main geojson class
@@ -79,6 +80,9 @@ class GeoJson {
   /// The stream indicating that the parsing is finished
   /// Use it to dispose the class if not needed anymore after parsing
   Stream<bool> get endSignal => _endSignalController.stream;
+
+  bool _disposed = false;
+
   // internal method
   Stream<T> _getGeoStream<T>() => _processedFeaturesController.stream
       .asBroadcastStream()
@@ -91,20 +95,8 @@ class GeoJson {
       bool verbose = false,
       GeoJsonQuery? query,
       bool disableStream = false}) async {
-    final file = File(path);
-    if (!file.existsSync()) {
-      throw FileSystemException("The file ${file.path} does not exist");
-    }
-    String data;
-    try {
-      data = await file.readAsString();
-    } catch (e) {
-      throw FileSystemException("Can not read file $e");
-    }
-    if (verbose) {
-      print("Parsing file ${file.path}");
-    }
-    await _parse(data,
+    await _parse(
+        fileName: path,
         nameProperty: nameProperty,
         verbose: verbose,
         query: query,
@@ -116,13 +108,18 @@ class GeoJson {
           {String? nameProperty,
           bool verbose = false,
           bool disableStream = false}) =>
-      _parse(data,
+      _parse(
+          data: data,
           nameProperty: nameProperty,
           verbose: verbose,
           disableStream: disableStream);
 
   void _pipeFeature(GeoJsonFeature<dynamic> data,
       {required bool disableStream}) {
+    if (_disposed) {
+      return;
+    }
+
     dynamic item;
     switch (data.type) {
       case GeoJsonFeatureType.point:
@@ -151,8 +148,10 @@ class GeoJson {
         break;
       case GeoJsonFeatureType.geometryCollection:
     }
-    if (!disableStream) {
-      if (item != null) _processedFeaturesController.sink.add(item);
+    if (!disableStream && !_processedFeaturesController.isClosed) {
+      if (item != null) {
+        _processedFeaturesController.sink.add(item);
+      }
       _processedFeaturesController.sink.add(data);
     }
     features.add(data);
@@ -179,8 +178,9 @@ class GeoJson {
     unawaited(_feats.close());
   }
 
-  Future<void> _parse(
-    String data, {
+  Future<void> _parse({
+    String? data,
+    String? fileName,
     required bool disableStream,
     required bool verbose,
     String? nameProperty,
@@ -200,9 +200,14 @@ class GeoJson {
       throw ParseErrorException("Can not parse geojson");
     });
     final dataToProcess = _DataToProcess(
-        data: data, nameProperty: nameProperty, verbose: verbose, query: query);
+        data: data,
+        fileName: fileName,
+        nameProperty: nameProperty,
+        verbose: verbose,
+        query: query);
     unawaited(iso.run(<dynamic>[dataToProcess]));
     await finished.future;
+    iso.dispose();
     _endSignalController.sink.add(true);
   }
 
@@ -228,7 +233,8 @@ class GeoJson {
       throw ArgumentError("Provide data or parse some to run a search");
     }
     if (data != null) {
-      await _parse(data,
+      await _parse(
+          data: data,
           nameProperty: nameProperty,
           verbose: verbose,
           query: query,
@@ -251,7 +257,7 @@ class GeoJson {
       if (data is GeoJsonPoint) {
         final point = data;
         foundPoints.add(point);
-        if (!disableStream) {
+        if (!disableStream && !_processedFeaturesController.isClosed) {
           _processedFeaturesController.sink.add(point);
         }
       } else {
@@ -265,6 +271,7 @@ class GeoJson {
         points: points, point: point, distance: distance, verbose: verbose);
     unawaited(iso.run(<dynamic>[dataToProcess]));
     await finished.future;
+    iso.dispose();
     return foundPoints;
   }
 
@@ -303,7 +310,7 @@ class GeoJson {
       if (data is GeoJsonPoint) {
         final point = data;
         foundPoints.add(point);
-        if (!disableStream) {
+        if (!disableStream && !_processedFeaturesController.isClosed) {
           _processedFeaturesController.sink.add(point);
         }
       } else {
@@ -317,6 +324,7 @@ class GeoJson {
         _GeoFenceToProcess(points: points, polygon: polygon, verbose: verbose);
     unawaited(iso.run(<dynamic>[dataToProcess]));
     await finished.future;
+    iso.dispose();
     return foundPoints;
   }
 
@@ -346,8 +354,71 @@ class GeoJson {
 
   /// Dispose the class when finished using it
   void dispose() {
+    _disposed = true;
     _processedFeaturesController.close();
     _endSignalController.close();
+  }
+
+  static bool _isOverlapping(
+      GeoBoundingBox boundingBox, GeoJsonFeature<dynamic>? feature) {
+    if (feature == null) {
+      return false;
+    }
+
+    final dynamic geometry = feature.geometry;
+    switch (feature.type) {
+      case GeoJsonFeatureType.geometryCollection:
+        final collection = geometry as GeoJsonGeometryCollection;
+        for (final geometry
+            in collection.geometries ?? <GeoJsonFeature<dynamic>?>[]) {
+          if (_isOverlapping(boundingBox, geometry)) {
+            return true;
+          }
+        }
+        return false;
+      case GeoJsonFeatureType.multipolygon:
+        final multiPolygon = geometry as GeoJsonMultiPolygon;
+        for (final polygon in multiPolygon.polygons) {
+          if (boundingBox
+              .isOverlapping(polygon.geoSeries.expand((e) => e.geoPoints))) {
+            return true;
+          }
+        }
+        return false;
+      case GeoJsonFeatureType.polygon:
+        final polygon = geometry as GeoJsonPolygon;
+        if (boundingBox
+            .isOverlapping(polygon.geoSeries.expand((e) => e.geoPoints))) {
+          return true;
+        }
+        return false;
+      case GeoJsonFeatureType.multiline:
+        final multiLine = geometry as GeoJsonMultiLine;
+        if (boundingBox.isOverlapping(
+            multiLine.lines.expand((e) => e.geoSerie?.geoPoints ?? []))) {
+          return true;
+        }
+        return false;
+      case GeoJsonFeatureType.line:
+        final line = geometry as GeoJsonLine;
+        if (boundingBox.isOverlapping(line.geoSerie?.geoPoints ?? [])) {
+          return true;
+        }
+        return false;
+      case GeoJsonFeatureType.multipoint:
+        final multiPoint = geometry as GeoJsonMultiPoint;
+        if (boundingBox.isOverlapping(multiPoint.geoSerie?.geoPoints ?? [])) {
+          return true;
+        }
+        return false;
+      case GeoJsonFeatureType.point:
+        final point = geometry as GeoJsonPoint;
+        if (boundingBox.isOverlapping([point.geoPoint])) {
+          return true;
+        }
+        return false;
+    }
+    return false;
   }
 
   static GeoJsonFeature<dynamic>? _processGeometry(
@@ -446,10 +517,26 @@ class GeoJson {
         throw ArgumentError.notNull();
       }
     }
-    final data = dataToProcess.data;
+
     final nameProperty = dataToProcess.nameProperty;
     final verbose = dataToProcess.verbose;
     final query = dataToProcess.query;
+
+    var data = dataToProcess.data;
+    if (data == null) {
+      final file = File(dataToProcess.fileName!);
+      if (!file.existsSync()) {
+        throw FileSystemException("The file ${file.path} does not exist");
+      }
+      try {
+        data = file.readAsStringSync();
+      } catch (e) {
+        throw FileSystemException("Can not read file $e");
+      }
+      if (verbose) {
+        print("Parsing file ${file.path}");
+      }
+    }
     final decoded = json.decode(data) as Map<String, dynamic>;
     final feats = decoded["features"] as List<dynamic>;
     for (final dFeature in feats) {
@@ -470,7 +557,7 @@ class GeoJson {
           if (nameProperty != null) {
             feature.geometry.name = properties![nameProperty];
           }
-          for (final geom in geometry["geometries"]) {
+          for (final geom in geometry["geometries"] as Iterable) {
             feature.geometry.add(_processGeometry(
                 geom as Map<String, dynamic>, properties, nameProperty));
           }
@@ -544,10 +631,21 @@ class GeoJson {
           continue;
         }
       }
+      if (query?.boundingBox != null) {
+        if (!_isOverlapping(query!.boundingBox!, feature)) {
+          if (verbose == true) {
+            print(
+                "Skipping out of bounds feature ${feature?.type} ${feature?.geometry?.name}");
+          }
+          continue;
+        }
+      }
       if (iso != null) {
         iso.send(feature);
       } else {
-        print("FEAT SINK $feature / ${feature?.type}");
+        if (verbose == true) {
+          print("FEAT SINK $feature / ${feature?.type}");
+        }
         sink?.add(feature);
       }
       if (verbose == true) {
@@ -596,12 +694,18 @@ class GeoJson {
 
 class _DataToProcess {
   _DataToProcess(
-      {required this.data,
+      {this.data,
+      this.fileName,
       required this.nameProperty,
       required this.verbose,
-      required this.query});
+      required this.query}) {
+    if (this.data == null && this.fileName == null) {
+      throw ArgumentError.notNull("Either data or filename should be provided");
+    }
+  }
 
-  final String data;
+  final String? fileName;
+  final String? data;
   final String? nameProperty;
   final bool verbose;
   final GeoJsonQuery? query;